requirements.txt
ansible-runner==2.3.2
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.19
lockfile==0.12.2
packaging==23.1
passlib==1.7.4
pexpect==4.8.0
ptyprocess==0.7.0
PyJWT==2.6.0
python-daemon==3.0.1
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3